create new folder structure, with budgetContainer separated from upcoming budgetFinderContainer
